task 4/requirements.txt
streamlit==1.24.0
python-dotenv
PyPDF2
openai
